modifier la strecture de controller livreur et vendeur et regeouper dans une controller admin
