+++ server/src/routes/AdminRoute/livreurRouter.ts
@@ -1,5 +1,6 @@
 import express from "express";
-import Livreur from "../../controllers/AdminController/livreurController";
+// import Livreur from "../../controllers/AdminController/livreurController";
+import adminController from "../../controllers/AdminController/ControllerAdmin";
 class LivreurRoute{
     public router:express.Router
 
@@ -9,11 +10,11 @@ class LivreurRoute{
     }
 
     private Livreurroute(){
-        this.router.post('/addlivreur',Livreur.AddLivreur)
-        this.router.get('/AfficheLivreur/:id',Livreur.AfficheLivreur)
-        this.router.get('/Alllivreur',Livreur.AllLivreur)
-        this.router.put('/updatlivreur/:id',Livreur.UpadatLivreur)
-        this.router.delete('/deletlivreur/:id',Livreur.DeleteLivreur)
+        this.router.post('/addlivreur',adminController.AddLivreur)
+        this.router.get('/AfficheLivreur/:id',adminController.AfficheLivreur)
+        this.router.get('/Alllivreur',adminController.AllLivreur)
+        this.router.put('/updatlivreur/:id',adminController.UpadatLivreur)
+        this.router.delete('/deletlivreur/:id',adminController.DeleteLivreur)
     }
 
 }
