+++ server/src/routes/AdminRoute/vendeurRouter.ts
@@ -1,5 +1,6 @@
 import express from "express";
-import vendeurController from "../../controllers/AdminController/vendeurController"
+// import vendeurController from "../../controllers/AdminController/vendeurController"
+import adminController from "../../controllers/AdminController/ControllerAdmin";
 class  VendeurRoute {
 	public router:express.Router; 
 	constructor(){
@@ -8,11 +9,11 @@ class  VendeurRoute {
 	}
 	
 	private Vendeurroute(){
-	this.router.post('/Addvendeur',vendeurController.Addvendeur)
-	this.router.put('/Updutevendeur/:id',vendeurController.Updutevendeur)
-	this.router.delete('/Deletevendeur/:id',vendeurController.Deletevendeur)
-	this.router.get('/Affichevendeur/:id',vendeurController.Affichevendeur)
-	this.router.get('/Allvendeur',vendeurController.Allvendeur)
+	this.router.post('/Addvendeur',adminController.Addvendeur)
+	this.router.put('/Updutevendeur/:id',adminController.Updutevendeur)
+	this.router.delete('/Deletevendeur/:id',adminController.Deletevendeur)
+	this.router.get('/Affichevendeur/:id',adminController.Affichevendeur)
+	this.router.get('/Allvendeur',adminController.Allvendeur)
 	}
 	
 }
